+++ changelog.md
@@ -2,6 +2,16 @@
 
 All notable changes to `laravel-alma-client` will be documented in this file.
 
+## **Version 1.0.2**
+- Final release with somme added methods
+  getPaymentStatus(string $paymentId): Payment
+  public function createPayment(array $data): Payment
+  getFeePlans(
+        string $kind = FeePlan::KIND_GENERAL,
+        $installmentsCounts = "all",
+        bool $includeDeferred = false
+    )
+
 ## **Version 1.0.1**
 
 ### Fixed